djangorestapi/requirements.txt
asgiref==3.2.3
bson==0.5.8
dataclasses==0.6
Django==2.2.9
django-mongodb-engine==0.6.0
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.11.0
djangotoolbox==1.8.0
djongo==1.3.0
mongoengine==0.19.1
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.2.4
